refactor: Fixture 위치 변경
