+++ infrastructure/mysql/src/test/kotlin/com/wespot/voteoption/fixture/VoteOptionJpaEntityFixture.kt
@@ -1,6 +1,7 @@
-package com.wespot.voteoption
+package com.wespot.voteoption.fixture
 
 import com.wespot.common.BaseEntity
+import com.wespot.voteoption.VoteOptionJpaEntity
 import java.time.LocalDateTime
 
 object VoteOptionJpaEntityFixture {
